PHP 8.4 | Tokenizer/PHP: fix union/intersection/DNF type tokenization with abstract properties
This commit adds support for `abstract` properties to the PHP Tokenizer as otherwise the retokenization of `|` to `T_TYPE_UNION`, `&` to `T_TYPE_INTERSECTION` and the parenthesis in DNF types to `T_TYPE_*_PARENTHESIS` would be broken.

Fixed now.

Includes tests.
Includes some extra tests for the `T_NULLABLE` tokenization as well, just to be on the safe side.