GH Actions: apply work-around for testing against PHP 8.5
Currently every single `composer install` on PHP 8.5 fails due to a deprecation notice coming from a Composer dependency during the reading of the `composer.json` file.

While it is expected that this will be fixed soonish (fix is already available in the dependency, we're just waiting for a new release of both the dependency as well as of Composer), I rather not have the red crosses on every single PR while working on getting PHPCS 4.0 released.

So, this commit adds a temporary work-around which should allow the tests to run against PHP 8.5 again.
This should also allow us to monitor if any of the other deprecation PRs which have gone into PHP 8.5 cause issues for PHPCS.

Once a new version of Composer has been released, we should be able to revert this commit.