Ruleset::processRule(): fix edge case bug - inconsistent handling of empty string array key
Unlikely scenario for real-life, but still a bug.

When an array property is set via the old comma-separated string ruleset format or via an inline `phpcs:set` annotation, and the key for an array element would be explicitly set to an empty string, the resulting array item in the property on the sniff class would be added without a key, resulting in a numeric key.

However, if the (not so) "new" ruleset format using `<element>` nodes is used and an array element key is an empty string, this would result in the array item being added with an empty string for the array key.

This is inconsistent behaviour and is now fixed.

As 2 out of 3 scenarios would result in a numeric key for PHPCS 3.x, the behaviour has now been standardized as such.

Note: for PHPCS 4.x, an empty string array key will remain an empty string and will not become a numeric key.
This is in-line with the type handling proposal in 708, which explicitly states:
> I do NOT intend to change the handling of array keys. These will remain strings at all times.