PHP 8.5 | Update tests involving the `disable_classes` ini setting
A number of tests verify the behaviour of PHPCS when a user tries to change a "php.ini only" ini setting, which should result in an error being displayed.

Unfortunately, the ini setting which was being used for these tests will be removed as of PHP 8.5, so this commit switches out the `disable_classes` ini setting for the `expose_php` ini setting to still be able to verify the behaviour.

Ref:
* https://wiki.php.net/rfc/deprecations_php_8_5#remove_disable_classes_ini_setting
* https://www.php.net/manual/en/ini.core.php