:fire: Hot Fix: remove FileList tests
PR 1113 introduced a set of tests for the `FileList` class to accompany a bug fix.

However, it now looks like these tests break the ability to run sniff tests for external standards using the PHPCS native test framework.
This is problematic and even more so as this will silently fail in CI, as PHPUnit 9 and lower exits with a `0` exit code when no tests are found.

For now, I'm removing the tests for the `FileList` class, while keeping the bugfix.

I've confirmed this fixes the issue by running the tests of two external standards which both use the PHPCS native test framework. Both were not running their tests with PHPCS 3.13.3 and the tests run correctly again with the `FileList` class tests removed.