PHP 8.5 | File::getMemberProperties(): fix "Using null as an array offset" deprecation
If a variable in the global scope is passed to `File::getMemberProperties()`, the token may not have any "conditions".
This would result in `array_keys()` returning an empty array, which will cause `array_pop()` to return `null`, leading to the deprecation notice.

Fixed now via some extra defensive coding.

This change is already covered via the existing tests.

Ref: https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_using_values_null_as_an_array_offset_and_when_calling_array_key_exists