PHP 8.5 | Fix another runtime deprecation notice
This fixes yet another "Using null as an array offset" deprecation notice on PHP 8.5. This one only occurred on PHPCS 4.0, not on 3.x.

The tests for this fix have been pulled in 1225 (to safeguard that the issue doesn't exist in 3.x) and merged up into the 4.x branch.

Note: as the `if` condition was pretty complex, I've done a small refactor to make the code more readable and understandable.