4.0 | Util classes: add parameter types
This commit explicitly does **not** include adding a parameter type to the `MessageCollector::add()` `$message` and `$type` parameters and the `Common::getSniffCode()` `$sniffsClass` parameter, as the pre-existing inline parameter validation is more comprehensive and more reliable.

Includes removing a type check from an "internal-only" class which is no longer necessary (and updating the associated test).