CHANGELOG: split into separate files per major
The changelog as it was, was by now > 8000 lines long and becoming unwieldy.

This commit splits the changelog file up into separate files per major release series.

It also adds the older changelog files to the `.gitattributes` file, so only the changelog files for the current and the previous major will be included in release zips.