Handle parse error in File::getMethodProperties()
While investigating the cause of a PHP Fatal error while processing some input
in the `PSR12.Functions.ReturnTypeDeclaration` sniff, I noticed that the
`File::getMethodProperties()` method did not return sensible values in for some
parse errors. The input file in question was
`src/Standards/Generic/Tests/Arrays/DisallowLongArraySyntaxUnitTest.2.inc`,
which contains an intentional parse error. The `File::getMethodProperties()`
method claimed that the return type of the `test` function was `<< HEAD`, which
is nonsense.

The change in this commit safeguards from such parse errors and resolves the
issue that was being encountered downstream in the sniff.